feat: Integrate create calendar and edit without auth
