+++ frontend/app/(tabs)/create.tsx
@@ -1,15 +1,104 @@
-import { View, Text, ScrollView, TextInput, Switch, Pressable, StyleSheet, useWindowDimensions } from "react-native";
+import { View, Text, TextInput, ScrollView, Pressable, StyleSheet, useWindowDimensions, Alert, ActivityIndicator } from "react-native";
 import { ThemedText } from "@/components/themed-text";
 import { Fonts } from "@/constants/theme";
 import { Ionicons } from "@expo/vector-icons";
 import { useState } from "react";
+import { useRouter } from "expo-router";
+import API_CONFIG from "@/constants/api";
+type PrivacyStatus = 'PRIVADO' | 'AMIGOS' | 'PUBLICO';
+type CalendarOrigin = 'CURRENT' | 'GOOGLE' | 'APPLE';
 
+interface PublishData {
+  nombre: string;
+  descripcion: string;
+  portada?: string;
+  estado: PrivacyStatus;
+  origen?: CalendarOrigin;
+}
 export default function CreateScreen() {
-  const [title, setTitle] = useState("");
-  const [description, setDescription] = useState("");
-  const [isPrivate, setIsPrivate] = useState(false);
+  const router = useRouter();
+  const [selectedPrivacy, setSelectedPrivacy] = useState<PrivacyStatus>('PRIVADO');
+  const [selectedOrigin, setSelectedOrigin] = useState<CalendarOrigin>('CURRENT');
+  const [isLoading, setIsLoading] = useState(false);
+  const [calendarData, setCalendarData] = useState<PublishData>({
+    nombre: "",
+    descripcion: "",
+    estado: 'PRIVADO',
+    origen: 'CURRENT',
+  });
+
   const { width } = useWindowDimensions();
   const isDesktop = width >= 768;
+
+  const privacyOptions: { label: string; value: PrivacyStatus; icon: string; description: string }[] = [
+    {
+      label: "Private",
+      value: "PRIVADO",
+      icon: "lock-closed-outline",
+      description: "Only you can see this calendar",
+    },
+    {
+      label: "Friends",
+      value: "AMIGOS",
+      icon: "people-outline",
+      description: "Visible to your friends only",
+    },
+    {
+      label: "Public",
+      value: "PUBLICO",
+      icon: "globe-outline",
+      description: "Visible to everyone",
+    },
+  ];
+
+  const originOptions: { label: string; value: CalendarOrigin }[] = [
+    { label: "Current", value: "CURRENT" },
+    { label: "Google", value: "GOOGLE" },
+    { label: "Apple", value: "APPLE" },
+  ];
+
+  const handlePublish = async () => {
+    if (!calendarData.nombre.trim()) {
+      Alert.alert("Error", "Calendar name is required.");
+      return;
+    }
+    setIsLoading(true);
+    try {
+      const response = await fetch(API_CONFIG.endpoints.createCalendar, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({
+          creador_id: 2, // TODO: replace with real user id when auth is done
+          nombre: calendarData.nombre,
+          descripcion: calendarData.descripcion,
+          estado: selectedPrivacy,
+          origen: selectedOrigin,
+        }),
+      });
+
+      if (!response.ok) {
+        const errorData = await response.json();
+        throw new Error(errorData.errors?.[0] ?? 'Unknown error');
+      }
+
+      router.replace('/(tabs)/calendars');
+    } catch (error) {
+      Alert.alert("Error", "Failed to publish calendar. Please try again.");
+      console.error("Publish error:", error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const handleDraft = async () => {
+    try {
+      // TODO: Save as draft
+      Alert.alert("Success", "Calendar saved as draft");
+      console.log("Calendar saved as draft");
+    } catch (error) {
+      Alert.alert("Error", "Failed to save draft");
+    }
+  };
 
   return (
     <View style={styles.wrapper}>
@@ -17,7 +106,7 @@ export default function CreateScreen() {
         contentContainerStyle={[styles.container, isDesktop && styles.containerDesktop]}
         showsVerticalScrollIndicator={false}
       >
-        {/* FORM CARD — centered on desktop */}
+        {/* FORM CARD */}
         <View style={[styles.card, isDesktop && styles.cardDesktop]}>
 
           {/* TITLE */}
@@ -27,73 +116,159 @@ export default function CreateScreen() {
             darkColor="#10464d"
             style={{ fontFamily: Fonts.rounded, textAlign: "center", marginVertical: 16 }}
           >
-            New Calendar
+            Create Calendar
           </ThemedText>
 
-          {/* COVER */}
-          <View style={styles.coverRow}>
-            <Text style={styles.label}>Cover:</Text>
-            <Pressable style={styles.coverBox}>
-              <Ionicons name="camera-outline" size={44} color="#aaa" />
-              <Text style={styles.coverHint}>Tap to upload</Text>
-            </Pressable>
+          {/* CALENDAR DETAILS */}
+          <View style={styles.inputSection}>
+            <Text style={styles.sectionTitle}>Calendar Details</Text>
+            <TextInput
+              style={styles.input}
+              placeholder="Calendar name"
+              placeholderTextColor="#aaa"
+              value={calendarData.nombre}
+              onChangeText={(text) => setCalendarData({ ...calendarData, nombre: text })}
+            />
+            <TextInput
+              style={[styles.input, styles.inputMultiline]}
+              placeholder="Description (optional)"
+              placeholderTextColor="#aaa"
+              value={calendarData.descripcion}
+              onChangeText={(text) => setCalendarData({ ...calendarData, descripcion: text })}
+              multiline
+              numberOfLines={3}
+            />
           </View>
 
-          {/* TITLE INPUT */}
-          <Text style={styles.label}>Title:</Text>
-          <TextInput
-            style={styles.input}
-            value={title}
-            onChangeText={setTitle}
-            placeholder="Give your calendar a name..."
-            placeholderTextColor="#bbb"
-          />
+          {/* DIVIDER */}
+          <View style={styles.divider} />
 
-          {/* DESCRIPTION INPUT */}
-          <Text style={styles.label}>Description:</Text>
-          <TextInput
-            style={[styles.input, styles.textArea]}
-            value={description}
-            onChangeText={setDescription}
-            placeholder="What is this calendar about?"
-            placeholderTextColor="#bbb"
-            multiline
-            numberOfLines={4}
-          />
-
-          {/* PRIVATE + FRIENDS ROW */}
-          <View style={styles.toggleRow}>
-            <View style={styles.toggleItem}>
-              <Text style={styles.toggleLabel}>Private</Text>
-              <Switch
-                value={isPrivate}
-                onValueChange={setIsPrivate}
-                trackColor={{ false: "#ccc", true: "#10464d" }}
-                thumbColor="#fff"
-              />
+          {/* CALENDAR SOURCE */}
+          <View style={styles.sourceSection}>
+            <Text style={styles.sectionTitle}>Calendar Source:</Text>
+            <View style={styles.originButtons}>
+              {originOptions.map((option) => (
+                <Pressable
+                  key={option.value}
+                  style={[
+                    styles.originBtn,
+                    selectedOrigin === option.value && styles.originBtnActive,
+                  ]}
+                  onPress={() => setSelectedOrigin(option.value)}
+                >
+                  <Text
+                    style={[
+                      styles.originBtnText,
+                      selectedOrigin === option.value && styles.originBtnTextActive,
+                    ]}
+                  >
+                    {option.label}
+                  </Text>
+                </Pressable>
+              ))}
             </View>
-
-            <Pressable style={styles.friendsButton}>
-              <Ionicons name="people-outline" size={16} color="#10464d" style={{ marginRight: 6 }} />
-              <Text style={styles.friendsButtonText}>Friends</Text>
-            </Pressable>
           </View>
 
-          {/* PUBLISH BUTTON — inside card on desktop, absolute on mobile */}
-          {isDesktop && (
-            <Pressable style={[styles.publishButton, { marginTop: 32 }]}>
-              <Text style={styles.publishText}>Publish</Text>
+          {/* DIVIDER */}
+          <View style={styles.divider} />
+          <View style={styles.privacySection}>
+            <Text style={styles.sectionTitle}>Who can see this?</Text>
+
+            {privacyOptions.map((option) => (
+              <Pressable
+                key={option.value}
+                style={[
+                  styles.privacyOption,
+                  selectedPrivacy === option.value && styles.privacyOptionSelected,
+                ]}
+                onPress={() => setSelectedPrivacy(option.value)}
+              >
+                <View style={styles.privacyIconRadius}>
+                  <Ionicons
+                    name={option.icon as any}
+                    size={20}
+                    color={selectedPrivacy === option.value ? "#10464d" : "#999"}
+                  />
+                </View>
+                <View style={styles.privacyContent}>
+                  <Text
+                    style={[
+                      styles.privacyLabel,
+                      selectedPrivacy === option.value && styles.privacyLabelSelected,
+                    ]}
+                  >
+                    {option.label}
+                  </Text>
+                  <Text style={styles.privacyDescription}>{option.description}</Text>
+                </View>
+
+                {/* Radio Button */}
+                <View
+                  style={[
+                    styles.radioButton,
+                    selectedPrivacy === option.value && styles.radioButtonSelected,
+                  ]}
+                >
+                  {selectedPrivacy === option.value && (
+                    <View style={styles.radioButtonInner} />
+                  )}
+                </View>
+              </Pressable>
+            ))}
+          </View>
+
+          {/* INFO BOX */}
+          <View style={styles.infoBox}>
+            <Ionicons name="information-circle-outline" size={20} color="#10464d" style={{ marginRight: 12 }} />
+            <Text style={styles.infoText}>
+              {selectedPrivacy === "PRIVADO"
+                ? "Only you can access and modify this calendar."
+                : selectedPrivacy === "AMIGOS"
+                ? "Your friends will receive an invitation to view this calendar."
+                : "Anyone with the link can view this calendar."}
+            </Text>
+          </View>
+
+          {/* ACTION BUTTONS */}
+          <View style={[styles.buttonGroup, isDesktop && styles.buttonGroupDesktop]}>
+            <Pressable
+              style={styles.draftButton}
+              onPress={handleDraft}
+            >
+              <Text style={styles.draftButtonText}>Save as Draft</Text>
             </Pressable>
-          )}
+
+            {isDesktop && (
+              <Pressable
+                style={[styles.publishButton, isLoading && styles.publishButtonDisabled]}
+                onPress={handlePublish}
+                disabled={isLoading}
+              >
+                {isLoading ? (
+                  <ActivityIndicator color="#fff" size="small" />
+                ) : (
+                  <Text style={styles.publishText}>Create Calendar</Text>
+                )}
+              </Pressable>
+            )}
+          </View>
 
         </View>
       </ScrollView>
 
-      {/* PUBLISH BUTTON — fixed at bottom on mobile only */}
+      {/* PUBLISH BUTTON — fixed at bottom on mobile */}
       {!isDesktop && (
         <View style={styles.publishContainer}>
-          <Pressable style={styles.publishButton}>
-            <Text style={styles.publishText}>Publish</Text>
+          <Pressable
+            style={[styles.publishButton, isLoading && styles.publishButtonDisabled]}
+            onPress={handlePublish}
+            disabled={isLoading}
+          >
+            {isLoading ? (
+              <ActivityIndicator color="#fff" size="small" />
+            ) : (
+              <Text style={styles.publishText}>Create Calendar</Text>
+            )}
           </Pressable>
         </View>
       )}
@@ -108,7 +283,7 @@ const styles = StyleSheet.create({
   },
   container: {
     paddingHorizontal: 24,
-    paddingBottom: 120,
+    paddingBottom: 140,
   },
   containerDesktop: {
     alignItems: "center",
@@ -116,7 +291,6 @@ const styles = StyleSheet.create({
     paddingBottom: 40,
   },
 
-  // CARD (wraps the form)
   card: {
     width: "100%",
   },
@@ -133,90 +307,171 @@ const styles = StyleSheet.create({
     elevation: 4,
   },
 
-  // COVER
-  coverRow: {
-    alignItems: "center",
+  // INPUT SECTION
+  inputSection: {
     marginBottom: 24,
-    marginTop: 4,
   },
-  coverBox: {
-    width: 140,
-    height: 140,
-    borderWidth: 2,
-    borderStyle: "dashed",
-    borderColor: "#aaa",
-    borderRadius: 12,
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#fff",
-    gap: 6,
-  },
-  coverHint: {
-    fontSize: 12,
-    color: "#aaa",
-  },
-
-  // INPUTS
-  label: {
-    fontSize: 13,
-    color: "#10464d",
-    fontWeight: "600",
-    marginBottom: 6,
-    marginTop: 16,
-  },
-  toggleLabel: {
+  sectionTitle: {
     fontSize: 15,
     color: "#10464d",
-    fontWeight: "600",
+    fontWeight: "700",
+    marginBottom: 12,
   },
   input: {
     borderWidth: 1.5,
-    borderColor: "#c0756a",
+    borderColor: "#e0e0e0",
     borderRadius: 10,
     paddingHorizontal: 14,
-    paddingVertical: 10,
-    fontSize: 15,
-    backgroundColor: "#fff",
+    paddingVertical: 12,
+    fontSize: 14,
     color: "#333",
+    backgroundColor: "#fff",
+    marginBottom: 12,
   },
-  textArea: {
-    height: 100,
+  inputMultiline: {
+    height: 90,
     textAlignVertical: "top",
   },
 
-  // TOGGLES
-  toggleRow: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    marginTop: 24,
-    backgroundColor: "#fff",
-    borderRadius: 12,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
+  divider: {
+    height: 1,
+    backgroundColor: "#e8e8e8",
+    marginVertical: 24,
   },
-  toggleItem: {
+
+  // SOURCE SECTION
+  sourceSection: {
+    marginBottom: 24,
+  },
+  originButtons: {
     flexDirection: "row",
-    alignItems: "center",
     gap: 10,
+    marginBottom: 24,
   },
-  friendsButton: {
-    flexDirection: "row",
-    alignItems: "center",
+  originBtn: {
+    flex: 1,
     borderWidth: 1.5,
+    borderColor: "#c0756a",
+    borderRadius: 10,
+    paddingVertical: 10,
+    alignItems: "center",
+    backgroundColor: "#fff",
+  },
+  originBtnActive: {
     borderColor: "#10464d",
-    borderRadius: 20,
-    paddingHorizontal: 16,
-    paddingVertical: 7,
     backgroundColor: "#f0f5f5",
   },
-  friendsButtonText: {
+  originBtnText: {
+    color: "#999",
+    fontSize: 13,
+    fontWeight: "600",
+  },
+  originBtnTextActive: {
     color: "#10464d",
+  },
+
+  // PRIVACY SECTION
+  privacySection: {
+    marginBottom: 24,
+  },
+  privacyOption: {
+    flexDirection: "row",
+    alignItems: "center",
+    padding: 14,
+    marginBottom: 10,
+    borderWidth: 1.5,
+    borderColor: "#e0e0e0",
+    borderRadius: 10,
+    backgroundColor: "#fff",
+  },
+  privacyOptionSelected: {
+    borderColor: "#10464d",
+    backgroundColor: "#f0f5f5",
+  },
+  privacyIconRadius: {
+    width: 44,
+    height: 44,
+    borderRadius: 22,
+    backgroundColor: "#f0f5f5",
+    alignItems: "center",
+    justifyContent: "center",
+    marginRight: 12,
+  },
+  privacyContent: {
+    flex: 1,
+  },
+  privacyLabel: {
     fontSize: 14,
+    color: "#333",
+    fontWeight: "600",
+    marginBottom: 2,
+  },
+  privacyLabelSelected: {
+    color: "#10464d",
+  },
+  privacyDescription: {
+    fontSize: 12,
+    color: "#999",
+  },
+  radioButton: {
+    width: 20,
+    height: 20,
+    borderRadius: 10,
+    borderWidth: 2,
+    borderColor: "#999",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  radioButtonSelected: {
+    borderColor: "#10464d",
+  },
+  radioButtonInner: {
+    width: 10,
+    height: 10,
+    borderRadius: 5,
+    backgroundColor: "#10464d",
+  },
+
+  // INFO BOX
+  infoBox: {
+    flexDirection: "row",
+    backgroundColor: "#f0f5f5",
+    borderRadius: 10,
+    padding: 12,
+    marginBottom: 24,
+    borderLeftWidth: 3,
+    borderLeftColor: "#10464d",
+  },
+  infoText: {
+    flex: 1,
+    fontSize: 12,
+    color: "#10464d",
+    lineHeight: 16,
+  },
+
+  // BUTTONS
+  buttonGroup: {
+    flexDirection: "row",
+    gap: 12,
+  },
+  buttonGroupDesktop: {
+    justifyContent: "space-between",
+  },
+  draftButton: {
+    flex: 1,
+    borderWidth: 1.5,
+    borderColor: "#10464d",
+    borderRadius: 10,
+    paddingVertical: 12,
+    alignItems: "center",
+    backgroundColor: "#fff",
+  },
+  draftButtonText: {
+    color: "#10464d",
+    fontSize: 15,
     fontWeight: "600",
   },
 
-  // PUBLISH
   publishContainer: {
     position: "absolute",
     bottom: 24,
@@ -224,6 +479,7 @@ const styles = StyleSheet.create({
     right: 24,
   },
   publishButton: {
+    flex: 1,
     backgroundColor: "#10464d",
     borderRadius: 30,
     paddingVertical: 16,
@@ -233,6 +489,9 @@ const styles = StyleSheet.create({
     shadowRadius: 8,
     shadowOffset: { width: 0, height: 4 },
     elevation: 6,
+  },
+  publishButtonDisabled: {
+    opacity: 0.6,
   },
   publishText: {
     color: "#fff",
